Remove -specs=...redhat-hardened... from Perl options
Red Hat's ExtUtils::Embed returns CLAGS & LDFLAGS that produce a
PIE position-independent executable. However libhts.a and the rest
of bcftools's *.o files are not compiled as PIC, so linking fails
as these other objects use relocations that are invalid for PIE.

While bcftools could link against libhts.so and compile the rest of
its objects with -fpic, the logistics are non-trivial. So it's easier
to omit the redhat-hardened-cc1 and redhat-hardened-ld specs that set
up for building a PIE executable. Fixes #1322.