Fix GT indexing
When setting value by determining index from the genotype, we face the problem
of how to interpret truncating arrays. Say we have TAG defined as Number=. and
     GT:TAG   1/1:0,1,2  0/0:0
Then when querying we expect the following expression to evaluate for the second
sample as
     -i 'TAG[1:1]="."'  .. true
     -i 'TAG[1:GT]="."' .. false
The problem is that the implementation truncates the number of fields, filling
usually fewer than the original number of per-sample values. This is fixed by
adding an exception that makes the code aware of this.

Fixes #2133