Add short option -W for --write-index.
Given --write-index now takes an optional argument, and optional long
options are --long-opt=arg and short options are -larg, I chose to
also accept -l=arg for the sake of consistency and ease of
documentation.  The standard -larg still works too.

If we wish to stick strictly to the normal conventions, then this is a
trivial change in version.c (and a search and replace in the
documentation).

Also fix formatting bug in bcftools merge man page section leading to
a lot of underlined text.

Fixes #2139