Revert "Support dynamic variables read from a tab-delimited annotation file"
This needs more work, didn't realize that the effect of the -c option is to
update the target VCF as well, so can't be used simply for matching. For
that, probably a new prefix will need to be introduced, but that's already
confusing as it is, for example ~ has double meaning.

This reverts commit e9bff3fb823e1fd47c3ca42b615b1eacb297028c. Switch to
branch bt-2151-annotate-ext to see this code.