Fix a trivial undefined behaviour sanitizer issue.
Memcpy strictly forbids passing in NULL as an argument, even when the
size of data to copy is 0.  It works on our systems and quite likely
on everybodies, and the compiler is highly unlikely to be able to
optimise this away ever as it's user-data driven.  However this is the
only place in the entire test harness that hits clang's ubsan,
clearing the way to potentially adding it to CI at some point.