sort: Consolidate temporary files to limit the number present
While sorting, merge temporary files when they become too numerous,
to keep the total count down.  Merging happens in layers of up
to 32 files.  Files in full layers are merged into the next layer
with a free slot.  As they move up, the merged files become
exponentially bigger, but are merged exponentially less frequently
so the overall complexity of the sort operation remains around
O(n * log n).  The number of layers is set at 12 - if these are
all filled then everything is simply merged into a single file at
the highest level (although this won't happen due to the number of
records that would be needed to get there).  This means the
number of temporary files can never exceed 384.

Temporary files are now opened in exclusive mode to eliminate
any risk of name clashes causing files to be overwritten.  They
are also compressed at level 1 to reduce the space needed to
store them.

To ensure the merge part of the sort is stable, the index of each
temporary file is stored in `idx` in its blk_t structure and used
in the min heap comparison function.