Make bcftools sort fully stable
libc qsort() isn't guaranteed to be stable, so add a comparison
tie-breaker based on the location in the input to ensure stability.
This can be done fairly easily by comparing the bcf1_t pointers
as most of the records are stored in a big block of memory created
with a single allocation.  Any records that aren't in this block,
because they were too big to fit in the remaining space, are
flagged by abusing the bcf1_t::errcode field so that the stable
comparison function can treat them correctly.