Fix a severe bug in merging of FORMAT fields
This affected tags with Number=R and Number=A values. For example, rows with high-coverage
FORMAT/AD values (bigger or equal to 128) would have been assigned to incorrect samples in
the manner of skipping every other value in the input file and filling in gibberish for the
rest.

The bug was introduced in version 1.19 by 752ce93, referencing it here in case this commit
did not manage to fix everything.

Fixes #2244