New experimental plugin +noise-profile and mpileup2 API
The plugin is intended with variant filtering of de novo mutations
in trios. It collects error profiles in unrelated samples and scores
sites by their noisiness.

This is an initial, backup commit. Only SNVs are supported in this
version.

This also brings the new mpileup2 interface. Currently only the
legacy mpileup functionality is functional, but the new API is
already useful in that it allows to use it with much less coding.