Fix inserted consensus defaults for --indels-cns mode.
The alphabet used by the alignment algorithm and elsewhere in the code
is 0,1,2,3,4 for ACGTN.  When building the consensus for an insertion
we use bcf_cgp_append_cons to add data to cons_ins.  This in turn uses
data in "ins" which has been set via "seq_nt16_int[bam_seqi(seq,pos)]",
turning bam bases 0-15 to 0-4 inclusive.

However the initialisation sets the insertions to "N".  Mostly this is
overwritten with actual aligned data, but where not it causes an
overflow in the edlib alphabet lookups.  Note the case where we don't
fill it out, leaving N, is caused by a maximum insertion length of
1024.

I'm unsure of what the effect is of increasing this limit, but it was
set to be an arbitrary value so we could use local static buffers
without dynamic memory allocation.  An unbounded indel size may cause
performance bottlenecks or present a trivial DoS attack vector by
including data with extreme insertion sizes.  For now the 1024 limit
is left intact.

Fixes #2249