Don't try to automatically determine good sites
Originally, the code would attempt to identify good sites with a heuristic,
but that did not always work well. A more robust way is to use a pre-computed
profile and allow the user to override. Thus in this commit the format changes,
good_site column is removed, and if one wishes to calculate variances from
a pre-selected sites, one can just subset the sites and run with -r.

Todo: read variance profile from a file