Fix a regression with -Ou format being the same as -Ob.
Commit 30dbc3c broke the -Ou option when a filename is given. So
`bcftools sort -Ou -o out.bcf in.vcf` is treated as -Ob instead.

Hence this was a regression in functionality that arrived with
bcftools 1.12.  The fix is simply to honour Ou vs Ob even when
filenames are specified, and only use the filename suffix when it's
incompatible with the data type specified.

Fixes #2263