Deal with an assert caused by a very short start exon
This fixes (imperfectly) a problem when the first exon is very short
and the code detects an overflow, see test/csq/ENSCAFT00000047742
     >chr9:104-110
     ATGTCAGGGCC
       eee.eee

This is not a real fix, the code would need a revamp to handle well general cases
like this. However, it is not real biology, so we don't care that much. This is
just so that the program does not crash on it.

Fixes #2313