Make `--merge both` and `--merge none` work with indel-only records
The code was making the incorrect assumption of always having a SNP record
available. However, that is not always the case and prevents the merging of
G>GT,T with G>GT.

This change reorders the output lines so that indels can come first, depending
on the order of files on the command line.

Hopefully resolves #2339