Work around Microsoft's UCRT-x64 pathname rewriting.
MINGW x64 works fine, but UCRT-x64 fails one of the setGT tests.
Specifically, `-n c:././.` gets silently rewritten before main() so
the argv element contains `-n c;.\.\.` instead.  This is a pain, and
arguably we should provide a less problematic CLI alternative for this
as c: was always going to trip up Windows boxes.

However for now we can work around it by not having it as a separate
argument and letting getopt do the tokenisation of argv into option
and optarg for us by removing the space.

This doesn't fix the bug obviously, but it makes it pass tests if
using the UCRT environment.