Add options `--samples-list` and `--samples-file`
This is to allow renaming samples from a list of samples on command line,
rather than from a file of sample names. Unfortunately, the existing
option `-s, --samples` conflicts with the rest of bcftools, therefore
this is added as

    -n, --samples-list LIST    New sample names given as a comma-separated list
    -N, --samples-file FILE    New sample names in a file, see the man page for details

The old option remains valid but is not advertised in the usage page.

Resolves #2383