Fix handling of skipped lines in `bcftools norm -c x`
Make normalize_line() report when it skips a line due to a ref
mismatch so that normalize_vcf() knows that it has to go on
to the next one.

Failing to do this caused a crash when the first input line
was skipped because normalize_vcf() attempted to access
data via args->rbuf when nothing had been added.

The input file for the `bcftools norm -c x` did not catch this
because the line it skipped was towards the end of the file.
An extra record has been added at the front, which does trigger
the issue on the original code.