Ignore first phasing bit in process_gt_to_hap / process_gt_to_hap2
HTSlib version 1.22 introduced prefixed phasing support for VCF
files with version 4.4 or later.  Prior to this the first GT phase
(when reading VCF) was always zero; after it is set either
explicitly or when all other alleles are phased.  This updates
process_gt_to_hap() and process_gt_to_hap2() to ignore the first
phase bit, removing an assumption it is always zero.

Fixes incorrect reporting of phase when using HTSlib 1.22 to read
VCF files with version 4.4 or 4.5.