Fix inadvertent variable length array
The C standard says that for an array to not be a variable length
type, it's size must be a constant integer expression.  A const
variable does not satisfy the definition of a constant integer
expression, but compliers may accept them as an extension.  We
used to get away with this, but the compiler shipped with MacOS
Tahoe is more strict and prints a warning.  Fix by changing the
const variable to a macro.

See https://stackoverflow.com/questions/72716896/ for more details.